various changes
- updateArch.sh script now runs the reflector commands
- added info section to readme
- modified buildme.sh to keep inline with upstream change:https://github.com/docker/docker/commit/ca94adc7c1ff7719be704282fe6daf1fb57161e0
- archive no longer left in /tmp


Former-commit-id: 9efe1e322790f71f5d0191141b111af591ad4b08

+++ buildme.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd -P)
 
@@ -9,26 +9,36 @@ cat > Dockerfile << EOF
 # https://github.com/l3iggs/docker-archlinux/blob/master/README.md
 FROM scratch
 MAINTAINER l3iggs <l3iggs@live.com>
-ADD archlinux.tar.xz /
-RUN pacman -Syyu --needed --noconfirm
 
-# install, run and remove reflector all in one line to prevent extra layer size
-RUN pacman -S --needed --noconfirm reflector; reflector --verbose -l 200 -p http --sort rate --save /etc/pacman.d/mirrorlist; pacman -Rs --noconfirm reflector
+# copy in the previously generated file system
+ADD archlinux.tar.xz /
+
+# update mirrorlist and packages
+ADD updateArch.sh /usr/sbin/updateArch.sh
+RUN chmod +x /usr/sbin/updateArch.sh; updateArch.sh
 EOF
 
 curl https://raw.githubusercontent.com/docker/docker/master/contrib/mkimage-arch.sh > /tmp/mkimage-arch.sh
 chmod +x /tmp/mkimage-arch.sh
 
-sed -i 's/| docker import - archlinux/-af archlinux.tar.xz/g' /tmp/mkimage-arch.sh
-sed -i '/docker run -t archlinux echo Success./d' /tmp/mkimage-arch.sh
+sed -i 's,| docker import - archlinux,-af /tmp/archlinux.tar.xz,g' /tmp/mkimage-arch.sh
+sed -i '/docker run --rm -t archlinux echo Success./d' /tmp/mkimage-arch.sh
+sed -i '$a chown '${USER}' /tmp/archlinux.tar.xz' /tmp/mkimage-arch.sh
 
 curl https://raw.githubusercontent.com/docker/docker/master/contrib/mkimage-arch-pacman.conf > /tmp/mkimage-arch-pacman.conf
 
 cd /tmp
 echo "Building Arch Linux-docker root filesystem archive."
-sudo /tmp/mkimage-arch.sh # this leaves /tmp/archlinux.tar.xz behind :-/
-echo "Arch Linux-docker root filesystem archive build complete."
-cp /tmp/archlinux.tar.xz ${DIR}/
+echo -e "\033[1msudo is required for the arch-chroot command.\033[0m"
+sudo /tmp/mkimage-arch.sh
+if [ -f /tmp/archlinux.tar.xz ]; then
+    echo "Arch Linux-docker root filesystem archive build complete!"
+    cp /tmp/archlinux.tar.xz ${DIR}/
+else
+    echo "The Arch Linux-docker root filesystem archive build failed."
+fi
 
-rm /tmp/mkimage-arch.sh
-rm /tmp/mkimage-arch-pacman.conf
+rm -rf /tmp/mkimage-arch.sh
+rm -rf /tmp/mkimage-arch-pacman.conf
+rm -rf /tmp/archlinux.tar.xz
+
